Organize js files into folders
